[boolToString] move, add logger
